Restructuring, add Nicole
I redid the format of the file so that main only reads in the data and creates the GUI.  All of the functions were moved to the library.pf file.  I also added Nicole's working per printer comparison to the library and added a button here to run that function.